[fix] get meetings details api 수정

+++ src/main/java/org/pingle/pingleserver/service/UserMeetingService.java
@@ -57,6 +57,8 @@ public class UserMeetingService {
     public Long cancelMeeting(Long userId, Long meetingId) {
         UserMeeting userMeeting = userMeetingRepository.findByUserIdAndMeetingId(userId, meetingId)
                 .orElseThrow(() -> new CustomException(ErrorMessage.RESOURCE_NOT_FOUND));
+        if(userMeetingRepository.existsByUserIdAndMeetingIdAndMeetingRole(userId, meetingId, MRole.OWNER))
+            throw new CustomException(ErrorMessage.BAD_REQUEST);
         userMeetingRepository.delete(userMeeting);
         return userMeeting.getId();
     }
